Move layout screen and pages to features folder
