+++ lib/features/pages/quizes/widgets/quiz_result_widget.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:gap/gap.dart';
 
-import '../../../../../core/config/texts/text_styles.dart';
-import '../../../../../core/config/theme/colors_manager.dart';
+import '../../../../core/config/texts/text_styles.dart';
+import '../../../../core/config/theme/colors_manager.dart';
 
 class QuizResultWidget extends StatelessWidget {
   final int score;
