+++ lib/features/pages/quizes/widgets/view_quiz_button.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
+import '../../../../core/helpers/extensions/extensions.dart';
+import '../../../quiz_details/quiz_details_screen.dart';
+
 class ViewQuizButton extends StatelessWidget {
   const ViewQuizButton({super.key});
 
@@ -9,6 +12,7 @@ class ViewQuizButton extends StatelessWidget {
     return GestureDetector(
       onTap: () {
         // TODO: navigate to Quiz Details Page
+        context.pushNamed(QuizDetailsScreen.routeName);
       },
       child: SvgPicture.asset("assets/icons/next_arrow_icon.svg"),
     );
